Project that uses promises
The project uses promises to make sure that data is available to a view before it moves to the view.

This project also works with the API setup for use with the production apps. E.g. Setting authentication for HTTP requests in app.run and using the service provided by the module.js file generated by the Restlet API.