Duplication of activities in event activity mappings
Made changes to the activity list controller to fix a duplication bug. Previously the user could select an activity even if it was already associated with the event through a mapping.
Now the activity list waits for a list of event to activity mapping objects for the event that is being edited is loaded, and when the user clicks on the activity it checks if that activity already exists by looping through the mappings for the event and checking to see if there is a match.
If there is a match then the user is presented with a popup alert that tells them that they've already added the activity to the event.